requirements.txt
aerich==0.8.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.8.0
asyncclick==8.1.8.0
asyncpg==0.30.0
bcrypt==4.3.0
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
dictdiffer==0.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.11
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
iso8601==2.1.0
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
psycopg2==2.9.10
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pypika-tortoise==0.5.0
pytest==8.3.5
pytest-asyncio==0.26.0
python-dotenv==1.0.1
pytz==2025.2
sniffio==1.3.1
starlette==0.46.1
tomlkit==0.13.2
tortoise-orm==0.24.2
typing-inspection==0.4.0
typing_extensions==4.12.2
uvicorn==0.34.0
